Update icon for client applications.
Adds several new explicit resolutions in the icon file for wider OS
support. This fixes display as a desktop icon by supplying the
expected size rather than relying upon scaling.

It does *not* compensate for the current bug in the Windows 10
taskbar, which uses the 32x32 icon but resizes it to 24x24 even
if a 24x24 image is available (which we now provide for
forward-compatibility).